# docs/swagger.yaml
basePath: /
definitions:
...
  database.Agent:
    properties:
      arch:
        type: string
      hostname:
        type: string
      id:
        type: string
      ip:
        type: string
      last_seen:
        type: string
      notes:
        type: string
      os:
        type: string
      registered_at:
        type: string
      tags:
        items:
          type: string
        type: array
      token:
        type: string
    type: object
  database.ErrorResponse:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
host: localhost:8443
info:
  contact:
    email: admin@example.com
    name: C2 Administrator
  description: API documentation for the C2 (Command and Control) server.
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
  title: C2 Server API
  version: "1.0"
paths:
  /admin/agents:
    get:
      description: Returns a list of all agents currently registered in the system
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/database.Agent'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/database.ErrorResponse'
      summary: Retrieve all registered agents
      tags:
      - admin
...
swagger: "2.0"